refactor(home/cli): migrate themes to flake inputs
Instead of fetching (and caching) specific versions of catppuccin theme
files with `builtins.fetchurl`, I added the various catppuccin theme
repos as flake inputs, then pass those inputs to the appropriate nix
modules.

This was motivated by a breaking change in the yazi theming API. When
fetching a file with `fetchurl`, nix will only verify that the resource
exists in the nix store, and will never invalidate stale files.