home/tldr: enable systemd service to update cache
This is currently enabled in addition to the module option to update on
invocation once a week.

The built-in update feature is annoying because it is blocking, and takes
a non-trivial amount of time to complete before outputting the requested
tldr document. This service should update the cache periodically in the
background, which should remove the need to wait for the cache to update
before giving the desired output.