feat(lib): update mkHomeConfig function
Add optional system arg (default to `x86_64-linux`), which is the only
system type I currently use. Use the system arg to select the correct
packages from an instantiated set of nixpkgs.

Additionally pass extra specialArgs to easily access the user and host
for which the config is built. This is useful when home-manager is used
standalone (which is the usecase for the `mkHomeConfig` function), since
an additional `nixConfig` special arg is not generated and passed to
modules.